+++ processor/src/test/resources/input/com/example/action/NoReportResultActionModel.java
@@ -1,0 +1,14 @@
+package com.example.action;
+
+import arez.annotations.Action;
+import arez.annotations.ArezComponent;
+
+@ArezComponent
+public abstract class NoReportResultActionModel
+{
+  @Action( reportResult = false )
+  public int doStuff()
+  {
+    return 0;
+  }
+}