+++ CHANGELOG.md
@@ -6,6 +6,13 @@
   exceed the size of an integer.
 * **\[core\]** In spy events, change the key used to identify the source of the event from source specific
   string (i.e. `computed`, `observable`, `observer`, etc.) to use a generic key `name`.
+* **\[core\]** Add flag `Flags.NO_REPORT_RESULT` that can be passed to `ArezContext.action(...)`,
+  `ArezContext.safeAction(...)` and `ArezContext.tracker(...)` methods that will result in the return value
+  (if any) from the action or observed function, not being specified in the `ActionCompletedEvent` spy event.
+* **\[core\]** Add the `reportResult` parameter to the `@Action` annotation so that the underlying actions
+  can be passed the `Flags.NO_REPORT_RESULT` flag.
+* **\[core\]** Add the `reportResult` parameter to the `@Observe` annotation so that the underlying observers
+  can be passed the `Flags.NO_REPORT_RESULT` flag.
 
 ### [v0.111](https://github.com/arez/arez/tree/v0.111) (2018-10-19)
 [Full Changelog](https://github.com/arez/arez/compare/v0.110...v0.111)